@@ -19,12 +19,12 @@ class ColumnStyle extends AbstractOptions
     /**
      * @var ColumnCondition[]
      */
-    protected $conditions = array();
+    protected $conditions = [];
 
     /**
      * @var Property[]
      */
-    protected $properties = array();
+    protected $properties = [];
 
     /**
      * @param  string $column
@@ -97,7 +97,7 @@ class ColumnStyle extends AbstractOptions
      */
     public function clearConditions()
     {
-        $this->conditions = array();
+        $this->conditions = [];
 
         return $this;
     }
@@ -154,7 +154,7 @@ class ColumnStyle extends AbstractOptions
      */
     public function clearProperties()
     {
-        $this->properties = array();
+        $this->properties = [];
 
         return $this;
     }
